Implement config.yml, now the app is fully customizable.

+++ src/main/java/me/laravieira/autolaraclone/resource/Resource.java
@@ -85,7 +85,7 @@ public class Resource {
         this.loader = loader;
     }
 
-    public Resource(int id, String name) {
+    public Resource(long id, String name) {
         this.id = id;
         this.name = name;
         this.file = new File("");
