+++ src/main/java/me/laravieira/autolaraclone/resource/Texture.java
@@ -2,7 +2,7 @@ package me.laravieira.autolaraclone.resource;
 
 public class Texture extends Resource {
 
-    public Texture(int id, String name) {
+    public Texture(long id, String name) {
         super(id, name);
     }
     public Texture(String name) {
